Added more file structure. Uses blank __init__.py file to treat each folder as a package.
