+++ fll_presenter/GUI/GUI_configurations/main_gui.py
@@ -2,19 +2,24 @@
 import tkinter
 import tkinter.colorchooser
 import platform
+import sys
+import time
 import customtkinter as ct
-import GUI_Elements.window_control as wc
-import GUI_Elements.timer as Timer
-import shared_functions
+from ..GUI_Elements import volume_control as vm
+from ..GUI_Elements import window_control as wc
+from ..GUI_Elements import timer as Timer
+import GUI.shared_functions as shared_functions
 
 system = platform.system()
+LASTCLICKMS = 0
+CLICKCOOLDOWN = 5           # Volume Control Cooldown In Seconds
 SCREENTOGGLE = False
 
 def run_gui():
-    """Function creates the timer gui and buttons as well as starting the event loop"""
+    """Function creates the main gui and buttons as well as starting the event loop"""
     controller = tkinter.Tk()
     controller.title("Controller Window")
-    controller.geometry("785x250")
+    controller.geometry("1325x250")
     controller.configure(background='#242424')
 
     timer_window = tkinter.Tk()
@@ -48,13 +53,42 @@ def run_gui():
     command=lambda: wc.toggle_window("Timer Window"),
     width=110, height=3, font=("Serif", 20))
 
+    volume_up_button = ct.CTkButton(controller, text="Volume Up",
+    command=lambda: shared_functions.call_volume_control(True, controller, system),
+    width=110, height=3, font=("Serif", 20))
+
+    volume_down_button = ct.CTkButton(controller, text="Volume Down",
+    command=lambda: shared_functions.call_volume_control(False, controller, system),
+    width=110, height=3, font=("Serif", 20))
+
+    min_volume_slider = ct.CTkSlider(controller, width=20, from_=0, to=100,
+    number_of_steps=100, orientation="vertical", command=shared_functions.set_volume_minimum)
+    min_volume_slider.set(vm.get_volume_control()[0])
+
+    max_volume_slider = ct.CTkSlider(controller, width=20, from_=0, to=100,
+    number_of_steps=100, orientation="vertical", command=shared_functions.set_volume_maximum)
+    max_volume_slider.set(vm.get_volume_control()[1])
+
+
     start_timer_button.pack(side=("left"), fill='y')
     stop_timer_button.pack(side=("left"), fill='y')
     timer_toggle_button.pack(side=("left"), fill='y')
+    volume_up_button.pack(side=("left"), fill='y')
+    volume_down_button.pack(side=("left"), fill='y')
+    min_volume_slider.pack(side=("left"), fill='y')
+    max_volume_slider.pack(side=("left"), fill='y')
+
 
     def set_timer_size(scale):
         """Function to update the size of the Timer widget"""
         Timer.update_size(timer,scale)
+
+    if system == "Windows":
+        application_selector = ct.CTkOptionMenu(controller, values=wc.get_window_executable_names(),
+        command=shared_functions.set_audio_application)
+        application_selector.pack(side="left")
+        application_selector.after(10000,
+        lambda:shared_functions.schedule_dropdown_update(controller, application_selector))
 
     timer_size = ct.CTkSlider(controller, width=20, from_=0, to=1000,
     orientation="vertical")
@@ -66,3 +100,6 @@ def run_gui():
     controller.protocol("WM_DELETE_WINDOW", lambda:shared_functions.kill_windows(controller,timer_window))
 
     timer_window.mainloop()
+
+
+
