+++ fll_presenter/GUI/__init__.py
@@ -1,0 +1,1 @@
+"""Creates a indicator to treat as a package. We can add code here later if needed."""
