+++ fll_presenter/GUI/config_gui.py
@@ -1,9 +1,9 @@
 """The module that creates and runs a configuration gui"""
 import platform
 import customtkinter as ct
-from GUI_configurations import main_gui
-import GUI_configurations.music_only_gui as music_gui
-import GUI_configurations.timer_only_gui as timer_gui
+from .GUI_configurations import main_gui
+from .GUI_configurations import music_only_gui as music_gui
+from .GUI_configurations import timer_only_gui as timer_gui
 
 system = platform.system()
 CONFIGWINDOW = None
